move query to separate directory
