+++ index.html
@@ -13,7 +13,7 @@
     <script src="/build/traitify.js"></script>
     <script>
       Traitify.setPublicKey("sptoechv411aqbqrp4l9a4eg2a");
-      Traitify.setHost("https://api.stag.awse.traitify.com");
+      Traitify.setHost("https://api.traitify.com");
 
       var widget;
